[refactor] abstract the project
