+++ src/main/java/com/oneswap/pubsub/RedisSubscriber.java
@@ -1,6 +1,8 @@
-package com.oneswap.service;
+package com.oneswap.pubsub;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.oneswap.service.LiquidityService;
+import com.oneswap.service.NetworkService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.redis.connection.Message;
