+++ src/main/java/com/oneswap/repositiry/impl/TransactionRepositoryImpl.java
@@ -1,6 +1,7 @@
-package com.oneswap.repositiry;
+package com.oneswap.repositiry.impl;
 
 import com.oneswap.model.Transaction;
+import com.oneswap.repositiry.TransactionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
